requirements.txt
# This can not be installed from pip, on Fedora you need to install rpm-python{,3}
#rpm-python
backports.lzma
copr
pyquery
six
GitPython
ansicolors
